requirements.txt
altair==4.2.2
pandas==1.4.2
Pillow==9.4.0
streamlit==1.18.1
